reactivate: Clean up /etc/schroot/schroot.conf.
If this update happens during a login session, then reactivate 1.x will
have modified schroot.conf, and schroot 1.3 will want to update that
file, and athena-auto-update is unlikely to deal with this correctly. To
avoid this, we clean up schroot.conf in our preinst; hopefully that
happens before schroot 1.3 gets unpacked. And you don't particularly
need to be in schroot.conf once you've logged in, anyway.

(reactivate 2 uses chroot.d to avoid this problem.)